# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shopline API - Local shop discovery, ordering and delivery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Phone already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/switch-role:
    post:
      operationId: switchRole
      tags: [auth]
      summary: Switch between buyer and seller mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchRoleBody"
      responses:
        "200":
          description: Role switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # SHOPS
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List orders for the current user (buyer or seller)
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a pickup order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (seller marks ready, buyer confirms pickup)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # DRIVERS
  /drivers/register:
    post:
      operationId: registerDriver
      tags: [drivers]
      summary: Register current user as a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDriverBody"
      responses:
        "201":
          description: Driver profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/me:
    get:
      operationId: getMyDriver
      tags: [drivers]
      summary: Get the authenticated user's driver profile
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: No driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDriverAvailability
      tags: [drivers]
      summary: Update driver availability status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [buyer, seller]
      required:
        - name
        - phone
        - password
        - role
    LoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
    SwitchRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, seller]
      required:
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [buyer, seller]
        activeRole:
          type: string
          enum: [buyer, seller]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - role
        - activeRole
        - createdAt

    # SHOP SCHEMAS
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        buyerId:
          type: integer
        shopId:
          type: integer
        shopName:
          type: string
        buyerName:
          type: string
        buyerPhone:
          type: string
        status:
          type: string
          enum: [pending, confirmed, ready, picked_up, cancelled]
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
        totalAmount:
          type: number
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - buyerId
        - shopId
        - shopName
        - buyerName
        - buyerPhone
        - status
        - fulfillmentType
        - totalAmount
        - items
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        imageUrl:
          type: ["string", "null"]
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
    CreateOrderBody:
      type: object
      properties:
        shopId:
          type: integer
        fulfillmentType:
          type: string
          enum: [pickup, delivery]
        note:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - shopId
        - fulfillmentType
        - items
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, ready, picked_up, cancelled]
      required:
        - status

    # REVIEW SCHEMAS
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleType:
          type: string
          enum: [bicycle, motorbike, car, van]
        vehicleDescription:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        totalDeliveries:
          type: integer
        rating:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - phone
        - vehicleType
        - isAvailable
        - totalDeliveries
        - createdAt
    RegisterDriverBody:
      type: object
      properties:
        vehicleType:
          type: string
          enum: [bicycle, motorbike, car, van]
        vehicleDescription:
          type: ["string", "null"]
      required:
        - vehicleType
    UpdateDriverBody:
      type: object
      properties:
        isAvailable:
          type: ["boolean", "null"]
        vehicleType:
          type: ["string", "null"]
        vehicleDescription:
          type: ["string", "null"]

    # DELIVERY SCHEMAS
...